fixed bugs and restructured file layout
