+++ .gitignore
@@ -1,4 +1,5 @@
 /**/dummy_main.py
+/**/.DS_Store
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
